upload multiple files concurrently new exported function (#101)
* upload multiple files concurrently new exported function

* re-arrange file.test.js test to have concurrently fcn in differernt test file

* fix commented out tests to use uploadFilesConcurrently, make requested changes
