+++ lib/file.js
@@ -244,7 +244,7 @@ async function uploadMultiPartFileConcurrently(filepath, target, options = {}) {
     if (!target) {
         throw new IllegalArgumentError('target not provided', target);
     } else if (!target.urls || target.urls.length === 0) {
-        throw new IllegalArgumentError('\'target.urls\' but be a non-empty array', target.urls);
+        throw new IllegalArgumentError('\'target.urls\' must be a non-empty array', target.urls);
     } else if (!isPositiveNumber(target.maxPartSize)) {
         throw new IllegalArgumentError('\'target.maxPartSize\' must be a positive number', target.maxPartSize);
     }
@@ -275,6 +275,81 @@ async function uploadMultiPartFileConcurrently(filepath, target, options = {}) {
         ],
         ...uploadOptions
     });
+}
+/**
+ * @typedef {Object} UploadAEMMultipartOptions
+ *
+ * @property {String} [method] Optional HTTP method (defaults to 'PUT')
+ * @property {Number} [timeout=30000] Optional socket timeout (ms)
+ * @property {Object} [headers] Optional override of request headers
+ * @property {Number} [retryMaxDuration=60000] time to retry until throwing an error (ms)
+ * @property {Number} [retryMaxCount] time to retry until throwing an error, overrides retryMaxDuration (ms)
+ * @property {Number} [retryInitialDelay=100] time between retries, used by exponential backoff (ms)
+ * @property {Boolean} [retryEnabled=true] retry on failure enabled
+ * @property {Boolean} [retryAllErrors=false] whether or not to retry on all http error codes or just >=500
+ * @property {Number} [preferredPartSize] Optional custom preferred part size. Might be adjusted depending on the target.
+ * @property {Number} [maxConcurrent] Optional concurrent amount.
+ */
+/**
+ * @typedef {Object} UploadAEMMultipartTarget
+ *
+ * @property {String[]} urls URLs
+ * @property {Number} [maxPartSize] Maximum size of each part
+ */
+/**
+ * @typedef {Object} UploadFile Object containing information about file to upload
+ *
+ * @param {UploadAEMMultipartTarget} target Target urls
+ * @param {String} filepath Path to locally saved file
+ */
+/**
+ * Upload multi part file using multiple simultaneous transfers
+ * Throws the first unrecoverable error if unsuccessful, all others are logged
+ * @param {UploadFile[]} files Array of UploadFile objects
+ * @param {UploadAEMMultipartOptions} options Additional options such as headers, retry features, etc 
+ * @returns {Promise} resolves when upload completes
+ */
+async function uploadFilesConcurrently(files, options = {}) {
+    const uploadFiles = [];
+    for (const file of files) {
+        const target = file.target;
+        const filepath = file.filepath;
+        // fail fast, keep error handling the same as older, non-concurrent function
+        if (!target) {
+            throw new IllegalArgumentError('target not provided', target);
+        } else if (!target.urls || target.urls.length === 0) {
+            throw new IllegalArgumentError('\'target.urls\' must be a non-empty array', target.urls);
+        } else if (!isPositiveNumber(target.maxPartSize)) {
+            throw new IllegalArgumentError('\'target.maxPartSize\' must be a positive number', target.maxPartSize);
+        }
+   
+        // Calculate the partSize based on the number of urls
+        const { size } = await getFileStats(filepath);
+        uploadFiles.push(
+            {
+                fileUrl: target.urls,
+                filePath: filepath,
+                fileSize: size,
+                maxPartSize: target.maxPartSize,
+                multipartHeaders: target.multipartHeaders
+            }
+        );
+    }
+    // to stay backwards compatible with older `uploadAEMMultipartFile` that uses `partSize`
+    options.preferredPartSize = options.partSize ? options.partSize : options.preferredPartSize;
+    
+    // extract upload options
+    const uploadOptions = filterObject(
+        options || {},
+        ['method', 'headers', ...CONCURRENCY_OPTIONS_FILTER,
+            ...RETRY_OPTIONS_FILTER, 'preferredPartSize']
+    );
+    const uploader = new BlockUpload();
+    return uploader.uploadFiles({
+        uploadFiles,
+        ...uploadOptions
+    });
+
 }
 
 /**
@@ -336,5 +411,6 @@ module.exports = {
     uploadFile,
     downloadFileConcurrently,
     uploadFileConcurrently,
-    uploadMultiPartFileConcurrently
+    uploadMultiPartFileConcurrently,
+    uploadFilesConcurrently
 };
