+++ lib/index.js
@@ -13,7 +13,7 @@
 'use strict';
 
 const { downloadStream, uploadStream, transferStream } = require('./stream');
-const { downloadFile, uploadFile, downloadFileConcurrently, uploadFileConcurrently, uploadMultiPartFileConcurrently } = require('./file');
+const { downloadFile, uploadFile, downloadFileConcurrently, uploadFileConcurrently, uploadMultiPartFileConcurrently, uploadFilesConcurrently } = require('./file');
 const { uploadAEMMultipartFile } = require('./aemmultipart');
 const { getResourceHeaders } = require('./headers');
 const { AEMUpload } = require("./aem/aemupload");
@@ -26,6 +26,7 @@ module.exports = {
     uploadFile,
     uploadFileConcurrently,
     uploadMultiPartFileConcurrently,
+    uploadFilesConcurrently,
     uploadAEMMultipartFile,
     getResourceHeaders,
     AEMUpload,
